requirements.txt
asgiref==3.8.1
Django==4.2
django-cors-headers==4.4.0
django-dotenv==1.4.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
PyJWT==2.9.0
pytz==2024.1
sqlparse==0.5.1
typing_extensions==4.12.2
